### 2022-01-05 #4034 @whedon
[PRE REVIEW]: APCtools: Descriptive and Model-based Age-Period-Cohort Analysis
**Submitting author:** @bauer-alex (<a href="http://orcid.org/0000-0003-3495-5131">Alexander Bauer</a>)
**Repository:** <a href="https://github.com/bauer-alex/APCtools" target ="_blank">https://github.com/bauer-alex/APCtools</a>
**Version:** 1.0.0
