### 2022-01-07 6cceaab
add CRAN downloads badge to readme
