### 2022-01-07 24497dc
changed license to MIT + added GitHub repo links to DESCRIPTION
