+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
-last_built: 2022-01-07T10:18Z
+last_built: 2022-01-07T10:34Z
 
