### 2022-01-07 #4034 @bauer-alex
Thanks @kthyng and @Nikoleta-v3 for your efforts!

Regarding the missing points raised by Kristen:

> @bauer-alex: please add an open source license and make sure one section in your paper is entitled "Statement of Need".

I now added both the open source license and the _Statement of Need_ section.
