### 2022-01-11 #4034 @Nikoleta-v3
👋 @allaway & @fernandomayer &  @franjoim & @allenzhuaz would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "APCtools: Descriptive and Model-based Age-Period-Cohort Analysis". You can find more information at the top of this Github issue 🆙
