### 2022-01-11 #4034 @allaway
@Nikoleta-v3 I am not an expert (or even really familiar with) Age-Period-Cohort analysis, but would be happy read up on it a bit and review.
