### 2022-01-17 6cbcc82
flexible colors in plot_1Dsmooth
