### 2022-02-03 0c42eff
fixed deprecated DESCRIPTION entry
