+++ DESCRIPTION
@@ -4,8 +4,7 @@ Version: 1.0.1
 Authors@R: c(
     person(given = "Alexander",  family = "Bauer",   email = "alexander.bauer@stat.uni-muenchen.de", role = c("aut", "cre"), comment = c(ORCID = "0000-0003-3495-5131")),
     person(given = "Maximilian", family = "Weigert", email = "maximilian.weigert@stat.uni-muenchen.de", role = "aut", comment = c(ORCID = "0000-0003-4400-134X")),
-    person(given = "Hawre",      family = "Jalal",   email = "hjalal@uottawa.ca", role = "aut", comment=c(ORCID = "0000-0002-8224-6834")),
-    person(given = "Pauline",    family = "Hohenemser", role = "ctb"))
+    person(given = "Hawre",      family = "Jalal",   email = "hjalal@uottawa.ca", role = "aut", comment=c(ORCID = "0000-0002-8224-6834")))
 Maintainer: Alexander Bauer <alexander.bauer@stat.uni-muenchen.de>
 Description: Age-Period-Cohort (APC) analyses are used to differentiate relevant drivers for long-term developments.
     The 'APCtools' package offers visualization techniques and general routines to simplify the workflow of an APC analysis.
