+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
-last_built: 2022-02-03T13:36Z
+last_built: 2022-02-11T07:07Z
 
