### 2022-02-15 bauer-alex/APCtools#4 @strengejacke
Statement of need missing in the readme
Related to https://github.com/openjournals/joss-reviews/issues/4056

According to the guidelines, the package website's readme should provide a statement of need. Please add such a section to your readme file.
