### 2022-02-15 bauer-alex/APCtools#5 @strengejacke
Community guidelines
Related to https://github.com/openjournals/joss-reviews/issues/4056

A passage or link to community guidelines is missing in the readme.
