### 2022-02-15 bauer-alex/APCtools#6 @strengejacke
Missing summary in software paper
Related to https://github.com/openjournals/joss-reviews/issues/4056

Please add a short summary to the paper, providing a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience.
