+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
-last_built: 2022-02-11T07:07Z
+last_built: 2022-02-16T09:58Z
 
