### 2022-02-16 bauer-alex/APCtools#5 @bauer-alex
We already had the section _Getting help_ in the readme to serve as such a passage.
We now named it _How to Contribute_, which is probably a clearer name, and slightly extended it to (i) link to the GitHub issue tracker and explicitly say (ii) that people can contact us via email and (iii) that we are very open for contributions, potentially via pull requests on GitHub.
