### 2022-02-16 bauer-alex/APCtools#6 @bauer-alex
We now edited the introduction of the paper accordingly, to outline the general problem and the functionality of the package in more high-level form.
I hope this fits your request?
