### 2022-02-16 #4056 @bauer-alex
@strengejacke Thanks so much for the review and the helpful comments in our issues.
We now changed the readme and the paper itself to try to resolve all issues (see [issue 1](https://github.com/bauer-alex/APCtools/issues/4), [issue 2](https://github.com/bauer-alex/APCtools/issues/5) and [issue 3](https://github.com/bauer-alex/APCtools/issues/6)).

Regarding the last of the above issues (the missing high-level summary) we were a bit unsure of what you think is missing in the paper, since we already had a short, high-level introduction. Either way, we now extended this high-level introduction a bit and restructured the initial part of the paper, s.t. it hopefully fits the need. What do you think?
