### 2022-02-16 #4056 @strengejacke
Yes, that looks better, thanks! I'm following the guidelines (see e.g. https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) that require the paper to have a short summary, followed by statement-of-need section. I'm not sure how @Nikoleta-v3 likes to handle this, if preferred as two separate sections or integrated in one (as you have it now), so I would leave this decision to the editors. In general, your additions would meet the needs for a short summary, you could just make two sections from your currently one statement-of-need section in the paper. Else, all remaining issues have been addressed.
