### 2022-02-16 #4056 @Nikoleta-v3
Thank you for your review @strengejacke!

I agree with the reviewer that the paper should include both a `summary` and a `statement-of-need`. The summary describes what the package is whereas the statement of need describes why the package is important. I find summaries very useful. You can just include 2-3 sentences describing `APCtools`. Here is an example: https://joss.theoj.org/papers/10.21105/joss.03465.

One comment regarding https://github.com/bauer-alex/APCtools/issues/5. I can see that your package includes tests.

Please include in your documentation guidelines on running these tests (apologies if they already exist and I missed them). That way if someone wants to contribute to the package they will know how to run your test suite.
