### 2022-02-17 #4056 @bauer-alex
I'll render the pdf once again, after another minor update to enlarge one figure in the rendered version.
