### 2022-02-20 #4056 @bauer-alex
@fernandomayer Do you know when you will be able to review the paper? Thing is that the paper is part of my PhD thesis which I want to submit in about 10 days. If possible, it would be great if you could do the revision next week?
I absolutely do not want to push you, as I know that this is just additional work on the side for you.
I just wanted to note that, if possible, it would be nice-to-have to start the revision in the following week. :)
If not possible, it's also no big deal for me to use the _in submission_ version of the paper for the thesis.
