### 2022-03-08 4a402c5
apply delta method transform only if CIs should be plotted
