### 2022-03-08 8dad71e
plot_smoothEffects: minor fix of ylim handling
