### 2022-03-27 #4034 @Nikoleta-v3
👋🏻  @allaway! I hope you are well 😄

A while back I asked if you could review this submission and you very kindly agreed.

Back then a few more people stepped up and I assigned different reviewers. Unfortunately, one of them is not available any more.

The review is currently taking place here:  https://github.com/openjournals/joss-reviews/issues/4056. As you can see one reviewer has already gone over the submission. I was wondering if you could help with the second review?

Thank you in advance!
