### 2022-03-28 #4034 @allaway
Hi @Nikoleta-v3, I am not able to review at the moment, but a colleague of mine (@jaybee84) is registered as a reviewer with JOSS and expressed interest in reviewing this article, in case you are open to considering her!
