### 2022-04-05 3f96efa
bug fix confidence interval for smooth effects
