+++ NEWS.md
@@ -1,3 +1,7 @@
+# APCtools 1.0.2 (2022-02)
+
+- Fixed computation of confidence intervals for smooth effects 
+
 
 # APCtools 1.0.1 (2022-01)
 
