### 2022-04-11 bauer-alex/APCtools#7 @jaybee84
License
Related to https://github.com/openjournals/joss-reviews/issues/4056

According to the guidelines, an [OSI approved](https://opensource.org/licenses/alphabetical) software license should be included in the LICENSE.txt file. The current LICENSE.txt file only contains copyright information. Can you please add a suitable license?
