### 2022-04-11 #4056 @jaybee84
@Nikoleta-v3
Re: Contribution and authorship: From the commit history of the package the contributions of 2 out of 3 authors is evident. From some of the references the contribution of the third author is implicit. Would it be ideal to include an `author contribution` section (similar to https://www.elsevier.com/authors/policies-and-guidelines/credit-author-statement) in the readme so that the contribution and attributions are more explicit?
