### 2022-04-12 48d0476
Bug fix in plot_densityMatrix, which wrongly assigned data to some cells in some situations
