+++ R/plot_densityMatrix.R
@@ -11,12 +11,13 @@
 #' @param dimensions Character vector specifying the two APC dimensions that
 #' should be visualized along the x-axis and y-axis. Defaults to
 #' \code{c("period","age")}.
-#' @param age_groups,period_groups,cohort_groups Each a list with one element
-#' specifying the borders of one row or column in the density matrix. E.g., if
-#' the period should be visualized in decade columns from 1980 to 2009, specify
+#' @param age_groups,period_groups,cohort_groups Each a list. Either containing
+#' purely scalar values or with each element specifying the two borders of one
+#' row or column in the density matrix. E.g., if the period should be visualized
+#' in decade columns from 1980 to 2009, specify
 #' \code{period_groups = list(c(1980,1989), c(1990,1999), c(2000,2009))}.
 #' The list can be named to specify labels for the categories. Only the two
-#' arguments should be passed that were specified by the \code{dimensions}
+#' arguments must be passed that were specified by the \code{dimensions}
 #' argument.
 #' @param highlight_diagonals Optional list to define diagonals in the density
 #' that should be highlighted with different colors. Each list element should be
@@ -49,7 +50,7 @@
 #'                      "2,000 - 6,000 km", "> 6,000 km")
 #' 
 #' age_groups    <- list(c(80,89),c(70,79),c(60,69),c(50,59),c(40,49),c(30,39),c(20,29))
-#' period_groups <- list(c(1971,1979),c(1980,1989),c(1990,1999),c(2000,2009),c(2010,2018))
+#' period_groups <- list(c(1970,1979),c(1980,1989),c(1990,1999),c(2000,2009),c(2010,2019))
 #' cohort_groups <- list(c(1980,1989),c(1970,1979),c(1960,1969),c(1950,1959),c(1940,1949),
 #'                       c(1930,1939),c(1920,1929))
 #' 
@@ -65,8 +66,7 @@
 #'                    y_var               = "mainTrip_distance",
 #'                    age_groups          = age_groups,
 #'                    period_groups       = period_groups,
-#'                    highlight_diagonals = list("born 1950 - 1959" = 8,
-#'                                               "born 1970 - 1979" = 10),
+#'                    highlight_diagonals = list(8, 10),
 #'                    log_scale           = TRUE)
 #' 
 #' # also mark different distance categories
@@ -77,8 +77,7 @@
 #'                    log_scale        = TRUE,
 #'                    y_var_cat_breaks = dist_cat_breaks,
 #'                    y_var_cat_labels = dist_cat_labels,
-#'                    highlight_diagonals = list("born 1950 - 1959" = 8,
-#'                                               "born 1970 - 1979" = 10),
+#'                    highlight_diagonals = list(8, 10),
 #'                    legend_title     = "Distance category")
 #' 
 #' # flexibly assign the APC dimensions to the x-axis and y-axis
@@ -99,16 +98,14 @@
 #'                    age_groups    = age_groups,
 #'                    period_groups = period_groups,
 #'                    log_scale     = TRUE,
-#'                    highlight_diagonals = list("born 1950 - 1959" = 8,
-#'                                               "born 1970 - 1979" = 10))
+#'                    highlight_diagonals = list(8, 10))
 #' 
 #' # plot categorical variables instead of metric ones
 #' plot_densityMatrix(dat                 = travel,
 #'                    y_var               = "household_size",
 #'                    age_groups          = age_groups,
 #'                    period_groups       = period_groups,
[…]
