+++ vignettes/main_functionality.Rmd
@@ -143,8 +143,8 @@ the arguments `age_groups`, `period_groups` and `cohort_groups`.
 ```{r, fig.width=8, message=FALSE}
 age_groups    <- list(c(80,89),c(70,79),c(60,69),c(50,59),
                       c(40,49),c(30,39),c(20,29))
-period_groups <- list(c(1971,1979),c(1980,1989),c(1990,1999),
-                      c(2000,2009),c(2010,2018))
+period_groups <- list(c(1970,1979),c(1980,1989),c(1990,1999),
+                      c(2000,2009),c(2010,2019))
 
 plot_densityMatrix(dat              = travel,
                    y_var            = "mainTrip_distance",
