### 2022-04-12 ac37dc7
paper: fixed the diagonal labels in the density matrix
