### 2022-04-12 #4056 @bauer-alex
Thanks so much for the input @jaybee84.
I have now added the explicit license text in the `LICENSE` file (see https://github.com/bauer-alex/APCtools/issues/7).

Also, I found that we made a minor mistake in the labels of the density matrix in the paper. I fixed it now and will re-generate the pdf with the next message.
Nothing else has changed in the paper.
