### 2022-04-19 bauer-alex/APCtools#8 @jaybee84
Minor additions to the manuscript
related to https://github.com/openjournals/joss-reviews/issues/4056

1. Statement of need : Please add in a bit more information about `identification problem` in the statement of need.
2. Is it possible to add a citation for `German Reiseanalyse survey– a repeated crosssectional study comprising information on German travelers between 1971 and 2018`
