+++ paper/paper.bib
@@ -23,8 +23,16 @@
   doi={10.1093/biostatistics/kxi028}
 }
 
+@misc{FUR_2022,
+  author={{Forschungsgemeinschaft Urlaub und Reisen e.V.}},
+  title={Survey of tourist demand in Germany for holiday travel and short breaks},
+  url={https://reiseanalyse.de/wp-content/uploads/2021/06/Reiseanalyse-2022_Informationsbroschuere.pdf},
+  year={2022},
+  note={accessed 20 April 2022}
+}
+
 @article{jalal_2020,
-  title={Hexamaps for Age-Period-Cohort Data Visualization and Implementation in R},
+  title={Hexamaps for Age-Period-Cohort Data Visualization and Implementation in {R}},
   author={Jalal, Hawre and Burke, Donald S},
   journal={Epidemiology (Cambridge, Mass.)},
   volume={31},
