### 2022-04-20 bauer-alex/APCtools#8 @bauer-alex
@jaybee84 We now added the citation for the Reiseanalyse (see the _Forschungsgemeinschaft Urlaub und Reisen e.V._ citation).

Regarding the statement of need, we added a few more sentences to make the identification problem a bit clearer.
The previous version was:
> The critical challenge in APC analysis is the linear dependency of the components age, period, and cohort (cohort = period - age). Flexible methods and visualization techniques are needed to circumvent this *identification problem*.

And the new version is:
> The critical challenge in APC analysis is to deal with the perfect linear dependency of the components age, period, and cohort (cohort = period - age). As of this *identification problem*, the estimation of a linear regression model with all three components as individual main effects is only possible when imposing additional constraints in the estimation process, like restricting one main effect to zero [@yang_land_2013]. Such explicit constraints, however, typically result in effect structures that are hard to interpret. Accordingly, flexible methods and visualization techniques are needed that rely on less restrictive assumptions to circumvent the identification problem.
