### 2022-04-20 #4056 @bauer-alex
I just updated the paper regarding the points raised by @jaybee84 in https://github.com/bauer-alex/APCtools/issues/8. See the respective issue for details.
