### 2022-04-21 bauer-alex/APCtools#8 @bauer-alex
Ah ok, makes sense.

Here our new proposal:
> The critical challenge in APC analysis is to deal with the perfect linear dependency of the components age, period, and cohort (cohort = period - age). `Due to this *identification problem*, inferring on the actual drivers behind observed temporal developments is difficult. For example, changes in recent years could be explained by developments related to the period, or by the fact that the respective observations only comprise later cohorts.` The estimation of a linear regression model with all three components as individual main effects is only possible when imposing additional constraints in the estimation process, like restricting one main effect to zero [@yang_land_2013]. Such explicit constraints, however, typically result in effect structures that are hard to interpret. Accordingly, flexible methods and visualization techniques are needed that rely on less restrictive assumptions to circumvent the identification problem.
