### 2022-04-21 bauer-alex/APCtools#8 @jaybee84
I just reopened the issue since the latest change has not been incorporated in the pdf article yet. Please feel free to close the issue once the new version of the document is generated.
