### 2022-04-21 #4056 @bauer-alex
The final points of @jaybee84 were now addressed in https://github.com/bauer-alex/APCtools/issues/8. I'm newly generating the pdf with the latest additions.
Thanks again for the input @jaybee84!

So I thing we resolved all issues, or am I missing something?
