### 2022-04-25 4d5f1fd
First attempt to build APChexamap with ggplot instead of using base R. Line-width, size and position of text still needs adjustements
