+++ R/ggplot_APChexamap.R
@@ -1,0 +1,401 @@
+#' Hexamap of an APC surface
+#' 
+#' Plot the heatmap of an APC structure using a hexagon-based plot with adapted
+#' axes. In this way, the one temporal dimension that is represented by the
+#' diagonal structure is visually not underrepresented compared to the other two
+#' dimensions on the x-axis and y-axis. \cr
+#' The function can be used in two ways: Either to plot the observed mean
+#' structure of a metric variable, by specifying \code{dat} and the variable
+#' \code{y_var}, or by specifying \code{dat} and the \code{model} object, to
+#' plot some mean structure represented by an estimated two-dimensional tensor
+#' product surface. The model must be estimated with \code{\link[mgcv]{gam}} or
+#' \code{\link[mgcv]{bam}}.
+#' 
+#' See also \code{\link{plot_APCheatmap}} to plot a regular heatmap.
+#' 
+#' If the plot is created based on the \code{model} object and the model was
+#' estimated with a log or logit link, the function automatically performs an
+#' exponential transformation of the effect.
+#' 
+#' @inheritParams plot_APCheatmap
+#' @param obs_interval Numeric specifying the interval width based on which the
+#' data is spaced. Only used if \code{y_var} is specified. Defaults to 1, i.e.
+#' observations each year.
+#' @param iso_interval Numeric specifying the interval width between the
+#' isolines along each axis. Defaults to 5.
+#' @param color_vec Optional character vector of color names, specifying the
+#' color continuum.
+#' @param color_range Optional numeric vector with two elements, specifying the
+#' ends of the color scale in the legend.
+#' @param line_width Line width of the isolines. Defaults to 0.5.
+#' @param line_color Character color name for the isolines. Defaults to gray.
+#' @param label_size Size of the labels along the axes. Defaults to 0.5.
+#' @param label_color Character color name for the labels along the axes.
+#' @param legend_title Optional character title for the legend.
+#' @param legend_limits Optional numeric vector passed as argument \code{limits}
+#' to \code{\link[ggplot2]{scale_fill_gradient2}}.
+#' 
+#' @return Plot created with \code{ggplot2}.
+#' 
+#' @import checkmate dplyr graphics ggplot2 reshape2
+#' @importFrom mgcv predict.gam
+#' @importFrom tidyr pivot_wider
+#' @export
+#' 
+#' @references Jalal, H., Burke, D. (2020). Hexamaps for Age–Period–Cohort
+#' Data Visualization and Implementation in R.
+#' \emph{Epidemiology}, 31 (6), e47-e49. doi: 10.1097/EDE.0000000000001236.
+#' 
+#' @author Hawre Jalal \email{hjalal@@pitt.edu},
+#' Alexander Bauer \email{alexander.bauer@@stat.uni-muenchen.de}
+#' Rickmer Schulte \email{rickmer.schulte@@stablab.stat.uni-muenchen.de}
+#' 
+#' @seealso \code{\link{plot_APCheatmap}}
+#' 
+#' @examples
+#' library(APCtools)
+#' library(mgcv)
+#' library(dplyr)
+#' 
+#' data(drug_deaths)
+#' 
+#' # restrict to data where the mortality rate is available
+#' drug_deaths <- drug_deaths %>% filter(!is.na(mortality_rate))
+#' 
+#' # hexamap of an observed structure
+#' ggplot_APChexamap(dat         = drug_deaths,
+#'                 y_var       = "mortality_rate",
+#'                 color_range = c(0,40))
+#' 
+#' # hexamap of a smoothed structure
+#' model <- gam(mortality_rate ~ te(age, period, bs = "ps", k = c(8,8)),
[…]
