### 2022-04-26 aced27e
First attempt to build APChexamap with ggplot instead of using base R.
Line-width, size and position of text still needs adjustments
