### 2022-05-03 #4056 @Nikoleta-v3
Thank you @bauer-alex for your work! Apologies for the delay I was out of the office 🇬🇷

Some last minor comments from me:

- Can you please fix the reference `Apc: Age-period-cohort analysis.` to be `apc: Age-period-cohort analysis.`
- In your README file it is easy to miss the documentation (useful materials). Could you make  `Useful materials` a section? and maybe rename it to `documentation and useful material`?
