### 2022-05-03 c48f89a
minor readme and paper bibliography update
