+++ paper/paper.bib
@@ -56,7 +56,7 @@
 }
 
 @Manual{R_apc,
-  title = {apc: Age-Period-Cohort Analysis},
+  title = {{apc: Age-Period-Cohort Analysis}},
   author = {Zoe Fannon and Bent Nielsen},
   year = {2020},
   note = {R package version 2.0.0},
