+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
-last_built: 2022-05-03T08:38Z
+last_built: 2022-05-03T12:08Z
 
