### 2022-05-03 #4056 @bauer-alex
Yeah, v1.0.3 is correct.

And yeah, the Ubuntu devel test currently fails. This randomly happens to me every once in a while with some packages on GitHub, even when no error in the codebase is present (-> all other OS tests run without any error). I hope this is not a problem? Otherwise I'll just re-run it in the following days until it works again, and then I'll let you know again.
